Working impl with service->resource dep injection, but need to redo api->service injection
