+++ src/main/java/tweeting/services/TwitterServiceComponent.java
@@ -1,0 +1,12 @@
+package tweeting.services;
+
+import dagger.Component;
+
+import javax.inject.Singleton;
+
+@Singleton
+@Component(modules = TwitterServiceModule.class)
+public interface TwitterServiceComponent {
+
+    TwitterService buildService();
+}
